+++ CHANGELOG.md
@@ -6,6 +6,8 @@ An EtherCAT master written in Rust.
 
 ## [Unreleased] - ReleaseDate
 
+## [0.3.2] - 2023-11-02
+
 ### Added
 
 - [#122] Added `Slave{Ref}::propagation_delay()` to get the EtherCAT propagation delay for a
@@ -203,8 +205,9 @@ An EtherCAT master written in Rust.
 - Initial release
 
 <!-- next-url -->
+[unreleased]: https://github.com/ethercrab-rs/ethercrab/compare/v0.3.2...HEAD
 
-[unreleased]: https://github.com/ethercrab-rs/ethercrab/compare/v0.3.0...HEAD
+[0.3.2]: https://github.com/ethercrab-rs/ethercrab/compare/v0.3.0...v0.3.2
 [0.3.1]: https://github.com/ethercrab-rs/ethercrab/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/ethercrab-rs/ethercrab/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/ethercrab-rs/ethercrab/compare/v0.2.0...v0.2.1